Create Dependency Loader Trait
