+++ plugin-name/includes/class-plugin-name.php
@@ -13,6 +13,8 @@ if ( ! class_exists( 'Plugin_Name' ) ) {
 
 	class Plugin_Name {
 
+		use Plugin_Name_Dependency_Loader;
+
 		/**
 		 *
 		 * @since    1.0.0
@@ -116,44 +118,6 @@ if ( ! class_exists( 'Plugin_Name' ) ) {
 			Plugin_Name_Actions_Filters::init_actions_filters();
 
 			$this->set_locale();
-		}
-
-		/**
-		 * Loads all Plugin dependencies
-		 *
-		 * @since    1.0.0
-		 */
-		private function load_dependencies( $class ) {
-
-			if ( false !== strpos( $class, Plugin_Name::CLASS_PREFIX ) ) {
-
-				$classFileName = 'class-' . str_replace( '_', '-', strtolower( $class ) ) . '.php';
-				$folder        = '/';
-
-				if ( false !== strpos( $class, '_Admin' ) ) {
-					$folder .= 'admin/';
-				}
-
-				if ( false !== strpos( $class, '_Public' ) ) {
-					$folder .= 'public/';
-				}
-
-				if ( false !== strpos( $class, Plugin_Name::CLASS_PREFIX . 'View' ) ) {
-					$path = Plugin_Name::get_plugin_path() . 'views' . $folder . $classFileName;
-					require_once( $path );
-				} elseif ( false !== strpos( $class, Plugin_Name::CLASS_PREFIX . 'Controller' ) ) {
-					$path = Plugin_Name::get_plugin_path() . 'controllers' . $folder . $classFileName;
-					require_once( $path );
-				} elseif ( false !== strpos( $class, Plugin_Name::CLASS_PREFIX . 'Model' ) ) {
-					$path = Plugin_Name::get_plugin_path() . 'models' . $folder . $classFileName;
-					require_once( $path );
-				} else {
-					$path = Plugin_Name::get_plugin_path() . 'includes/' . $classFileName;
-					require_once( $path );
-				}
-
-			}
-
 		}
 
 		/**
