+++ plugin-name/plugin-name.php
@@ -102,6 +102,7 @@ function run_plugin_name() {
 		 * The core plugin class that is used to define internationalization,
 		 * admin-specific hooks, and public-facing site hooks.
 		 */
+		require_once plugin_dir_path( __FILE__ ) . 'includes/trait-plugin-name-dependency-loader.php';
 		require_once plugin_dir_path( __FILE__ ) . 'includes/class-plugin-name.php';
 
 		/**
